Error handling

+++ main.py
@@ -44,7 +44,6 @@ def init_powerpoint():
 @eel.expose
 def start(input_file_name, output_file_name, send):    
     print(time.time())
-    print(send)
 
     from PPTX_to_PDF import main
     from sending import login, send_email
@@ -56,31 +55,36 @@ def start(input_file_name, output_file_name, send):
     import os
 
     if send:
-            try:
-                os.environ["MY_ADRESS"]
-            except KeyError:
-                eel.raise_error("Почтовые данные не найдены, но сертификаты созданы", "error")
+        try:
+            os.environ["MY_ADRESS"]
+        except KeyError:
+            eel.raise_error("Почтовые данные не найдены, но сертификаты созданы")
+            send = False
 
     data = all_names(input_file_name, output_file_name)
-    # То что ниже нужно для тестинга
 
-    os.makedirs(f"GENERATED_PPTX/{data[0]['date']}",
-                exist_ok=True)
-    os.makedirs(f"GENERATED_PDF/{data[0]['date']}",
-                exist_ok=True)
+    # Если Excel файл не найден
+    if data == 'Excel':
+        eel.raise_error('Excel файл не найден')
+        return
+
+    os.makedirs(f"GENERATED_PPTX/{data[0]['date']}", exist_ok=True)
+    os.makedirs(f"GENERATED_PDF/{data[0]['date']}", exist_ok=True)
 
     shutil.rmtree(f"GENERATED_PPTX/{data[0]['date']}")
     shutil.rmtree(f"GENERATED_PDF/{data[0]['date']}")
 
-    os.makedirs(f"GENERATED_PPTX/{data[0]['date']}",
-                exist_ok=True)
-    os.makedirs(f"GENERATED_PDF/{data[0]['date']}",
-                exist_ok=True)
+    os.makedirs(f"GENERATED_PPTX/{data[0]['date']}", exist_ok=True)
+    os.makedirs(f"GENERATED_PDF/{data[0]['date']}", exist_ok=True)
 
     powerpoint = init_powerpoint()
 
     for loc in data:
         file_name = PPTX_GENERATOR(loc)
+        if file_name == "Template":
+            eel.raise_error("Файл-шаблон не найден")
+            powerpoint.Quit()
+            return
         # command = "python PPTX_to_PDF.py " + file_name + " " + loc['date']
         main(file_name, loc['date'], powerpoint)
 
@@ -89,14 +93,15 @@ def start(input_file_name, output_file_name, send):
                 smtps = login()
                 send_email(loc['email'], smtps, loc['date'], file_name)
             except SMTPAuthenticationError:
-                eel.raise_error("Не верно указан пароль от почты.\n Статья о других возможных проблемах по ссылке", "error")
+                eel.raise_error("Не верно указан пароль от почты.\n Статья о других возможных проблемах по ссылке")
                 send = False
             except KeyError:
-                eel.raise_error("В Excel документе нет поля email, сертификаты не были отправлены", "error")
+                eel.raise_error("В Excel документе нет поля email, сертификаты не были отправлены")
                 send = False
         # file = file_name.replace("©", " ")
 
     powerpoint.Quit()
+    eel.raise_error("Процесс успешно завершен")
     print(time.time())
 
 if __name__ == "__main__":
